2018/07/29
2401 revised

+++ BOJ/kmp/2401.cpp11.cpp
@@ -36,7 +36,7 @@ void kmp(int itr)
         {
             matched ++;
 
-            if(matched == l[itr].size())
+            if(matched == (int)l[itr].size())
             {
                 w[i][itr] = true;
                 matched = p[matched-1];
